Updated README.md
It was thanks to Froren & Kris that it is possible to sort by an individual agent or agent organization. Made the changes to reflect this in README so it can help anybody else who might need it in the future. Special thanks to Froren for writing this wrapper. (EDIT: added Kristophor's name)